Distribution.Helper

Documentation

data Programs

Paths or names of various programs we need.

Constructors

Programs

Fields

cabalProgram :: FilePath

ghcProgram :: FilePath

ghcPkgProgram :: FilePath

defaultPrograms :: Programs

data QueryEnv

qeReadProcess :: QueryEnv -> FilePath -> [String] -> String -> IO String

How to start the cabal-helper process. Useful if you need to capture stderr output from the helper.

qePrograms :: QueryEnv -> Programs

qeProjectDir :: QueryEnv -> FilePath

Path to project directory, i.e. the one containing the project.cabal file

qeDistDir :: QueryEnv -> FilePath

Path to the dist/ directory

qeCabalPkgDb :: QueryEnv -> Maybe FilePath

Where to look for the Cabal library when linking the helper

qeCabalVer :: QueryEnv -> Maybe Version

If dist/setup-config wasn't written by this version of Cabal throw an error

defaultQueryEnv

Arguments

:: FilePath
Path to project directory, i.e. the one containing the project.cabal file

-> FilePath
Path to the dist/ directory

-> QueryEnv

Running Queries

data Query m a

Caches helper executable result so it doesn't have to be run more than once as reading in Cabal's LocalBuildInfo datatype from disk is very slow but running all possible queries against it at once is cheap.

runQuery :: Monad m => QueryEnv -> Query m a -> m a

runQuery query distdir. Run a Query. distdir is where Cabal's setup-config file is located.

Queries against Cabal's on disk state

packageDbStack :: MonadIO m => Query m [ChPkgDb]

List of package databases to use.

entrypoints :: MonadIO m => Query m [(ChComponentName, ChEntrypoint)]

Modules or files Cabal would have the compiler build directly. Can be used to compute the home module closure for a component.

sourceDirs :: MonadIO m => Query m [(ChComponentName, [FilePath])]

A component's source-dirs field, beware as if this is empty implicit behaviour in GHC kicks in.

ghcOptions :: MonadIO m => Query m [(ChComponentName, [String])]

All options cabal would pass to GHC.

ghcSrcOptions :: MonadIO m => Query m [(ChComponentName, [String])]

Only search path related GHC options.

ghcPkgOptions :: MonadIO m => Query m [(ChComponentName, [String])]

Only package related GHC options, sufficient for things don't need to access any home modules.

ghcMergedPkgOptions :: MonadIO m => Query m [String]

Like ghcPkgOptions but for the whole package not just one component

ghcLangOptions :: MonadIO m => Query m [(ChComponentName, [String])]

Only language related options, i.e. -XSomeExtension

pkgLicenses :: MonadIO m => Query m [(String, [(String, Version)])]

Get the licenses of the packages the current project is linking against.

Result types

newtype ChModuleName

Constructors

ChModuleName String

data ChComponentName

Constructors

ChSetupHsName

ChLibName

ChExeName String

ChTestName String

ChBenchName String

data ChPkgDb

Constructors

ChPkgGlobal

ChPkgUser

ChPkgSpecific FilePath

data ChEntrypoint

Constructors

ChSetupEntrypoint
Almost like ChExeEntrypoint but main-is could either be "Setup.hs" or "Setup.lhs". Since we don't know where the source directory is you have to find these files.

ChLibEntrypoint

Fields

chExposedModules :: [ChModuleName]

chOtherModules :: [ChModuleName]

ChExeEntrypoint

Fields

chMainIs :: FilePath

chOtherModules :: [ChModuleName]

General information

buildPlatform :: (FilePath -> [String] -> String -> IO String) -> IO String

Stuff that cabal-install really should export

getSandboxPkgDb

Arguments

:: (FilePath -> [String] -> String -> IO String)

-> FilePath
Cabal build platform, i.e. buildPlatform

-> Version
GHC version (cProjectVersion is your friend)

-> IO (Maybe FilePath)

Get the path to the sandbox package-db in a project

Managing dist/

prepare

Arguments

:: MonadIO m

=> (FilePath -> [String] -> String -> IO String)

-> FilePath
Path to project directory, i.e. the one containing the project.cabal file

-> FilePath
Path to the dist/ directory

-> m ()

Make sure the appropriate helper executable for the given project is installed and ready to run queries.

reconfigure

Arguments

:: MonadIO m

=> (FilePath -> [String] -> String -> IO String)

-> Programs
Program paths

-> [String]
Command line arguments to be passed to cabal

-> m ()

Run cabal configure

writeAutogenFiles

Arguments

:: MonadIO m

=> (FilePath -> [String] -> String -> IO String)

-> FilePath
Path to project directory, i.e. the one containing the project.cabal file

-> FilePath
Path to the dist/ directory

-> m ()

Create cabal_macros.h and Paths_<pkg> possibly other generated files in the usual place.

$libexec related error handling

data LibexecNotFoundError

This exception is thrown by all runQuery functions if the internal wrapper executable cannot be found. You may catch this and present the user an appropriate error message however the default is to print libexecNotFoundError.

Constructors

LibexecNotFoundError String FilePath

libexecNotFoundError

Arguments

:: String
Name of the executable we were trying to find

-> FilePath
Path to $libexecdir

-> String
URL the user will be directed towards to report a bug.

-> String